minihack: Change GoalGenerator to RewardManager
Instead of using the GoalGenerator class which was limited to sparse
reward goals and no customisation of the reward value, introduce a
RewardManager class which handles rewards and environment termination
for MiniHack.

This class enables creating events with associated
rewards, and specifying which events are required or sufficient for
completion. You can also specify which events can be repeated (for
example if you have many apples in the room, you could just add one
event which is repeatable to eat apples).

Due to funky imports (MiniHack and RewardManager depend on each other)
we only conditionally import MiniHack in reward_manager.py if we're type
checking, and use __future__.annotations to allows typing with the class
rather than the string.

Change all the MiniHackSkillEnv tasks to use the RewardManager (which
was easy).

Add a coordinate-based event which looks at the agent's coordinates for
giving reward.

Add SequentialRewardManager for enforcing a sequence of goals, and
GroupedRewardManager for arbitrary groupings of goals (i.e. complete all
tasks in group 1 or any task in group 2).

Override `step` in MiniHack to save the previous observation and action
so that they can used in _reward_fn and _is_episode_end (as that's where
we call the RewardManager). I've created an issue on the NLE repository
to change `_reward_fn` to actually take in the action, which would
remove the need for one of these hacks.

We still terminate if we reach a staircase down with positive reward,
which enables the adding of purely auxiliary rewards to any task (even
navigation tasks) - i.e. to eat as many apples as possible and then
reach the staircase. Of course if you don't put a staircase down in the
level then this isn't a problem.