minihack: Improve NetHackWiki integration
Add a method to the base MiniHack class which returns the list of page
contents (in the same format as the list of descriptions) for the
objects in the surrounding tiles of the agent. This uses some NLP
preprocessing to get the text into a format which is more likely to find
the correct wiki page (although there are probably still failured
cases).

The NLP preprocessin is optional - currently it only works if you've
installed the required packages.