minihack: Add deep exploration task
This task has an easy solution (go to the down staircase in the starting
room) and a more difficult solution (find the room with the apples, eat
them, and then go down the staircase. The optimal policy is to do the
second solution, as the apples provide positive reward, but it's likely
that the policy will converge on going down the stair-case to avoid
per-time-step negative rewards. Hence, an algorithm that can solve this
environment should exhibit deep exploration, which often goes counter to
policy optimsation.

The environment builds off the corridor environments, but with adjust
action space, reward function and added apples.