agent-rllib: Make script runnable with multiple algos
We (initially) support IMPALA, PPO and DQN. SAC with dictionary
observations is broken in a frustrating way which isn't easy to fix, so
I'm leaving that for now, but we have DQN as an off-policy algorithm.

The configuration should be fairly flexible. I'm planning to run some
experiments for MiniHack and for the Nethack Competition with these
baselines in the next few days.