minihack: Adjust deep explore to always end on staircase
For some reason having multiple down staircase results in one of them
not triggering the end of the episode. Hence, we remove the staircase in
the initial room, so that there's only one. The initial room wasn't
going to be the initial room anyway as staircase up has no effect.

Also add some possible messages to the eat task to catch rotten food or
finishing off partially eaten food.