minihack: Add seeds argument and functionality
Add the ability for users to specfifying the seeds used to generate the
levels. This allows tighter control over the generation process, and the
ability to test generalisation to levels that are guaranteed to be
unseen.

The `env.seed` method could be use already if desired, but it would
require a more manual approach which wraps the environments and handles
the seeding. This method will allow users to easily specify the seeds to
be used, giving a clean train/test split.