minihack: Add optional reward shaping to boxoban
The learning curves for these environments are flat. To alleviate this,
add a reward shaping bonus based on how many fountains are covered by
boulders. The reward is the difference in number of fountains covered by
boulders between the previous and current time-step, and so isn't
deceptive as the agent is penalised for pushing the boulder off the
fountain as much as it is for pushing it onto the fountain; hence the
agent won't just cycle between these two actions to get reward. The
coefficient of the reward shaping can be adjusted through an argument,
but is currently set to 0.1, which means the total reward will be 1.4
instead of 1.