minihack: Add mazeexplore task
This is a simpler (hopefully) version of the deep exploration task. The
agent starts at the top of the level, and must navigate a maze to the
exit in the middle of the level. However, there are apples in aroom at
the bottom of the level which provide high reward ife ater. Hence the
optimal strategy is to explore the maze deeply (regardless of whether
the exit has been found) until you find the apple room, and then eat the
apples and exit the level.

We provide an easy and a hard version, with the hard version being
larger and having more apples.