agent-rllib: Ensure full hydra config is used
Before, there were values in the hydra config (such as `lr`) which were
never being passed to RLlib, which is obviously not good. Also, it was
hard to get visibility in wandb for all the configuration options
passed, as they were too deep in the yaml hierarchy and were OmegaConf
objects rather than plain dicts.

To fix this, we update the default RLlib algorithm config with the hydra
config, and the algo-specific config, and then specific ways the hydra
config should be handled (i.e.  not just directly copying keys to
values, e.g. the model specification and number of actors/cpus).

To ensure RLlib is fine with this, we "create" a new trainer class with
this config as the default one, so we don't receive complaints about
unknown configuration keys.

THis ran for a few iterations without erroring on my computer, and so
should be working correctly, and taking in arguments passed in the
configuration or the command line as we expect.